// core/src/shared/git-provider/types.ts
export type GitProviderType = "gitea" | "github" | "gitlab";

/** Git Provider 模块配置选项 */
export interface GitProviderModuleOptions {
  /** Git Provider 平台类型 */
  provider: GitProviderType;
  /** 服务器 URL */
  baseUrl: string;
  /** API Token */
  token: string;
}

/** Git Provider 模块异步配置选项 */
export interface GitProviderModuleAsyncOptions {
  useFactory: (...args: unknown[]) => Promise<GitProviderModuleOptions> | GitProviderModuleOptions;
  inject?: any[];
}

/** 注入令牌 */
export const GIT_PROVIDER_MODULE_OPTIONS = "GIT_PROVIDER_MODULE_OPTIONS";